Fix broken line length in xt_words
xt_words didn't respect MAX_LINE_LENGTH because at the end of each line
it set the length counter to 0. Instead it should be set to the word
about to be printed because that's already in memory so its length
needs to be taken into consideration. This commit fixes that behaviour
and also adds one extra character to the count to account for the added
space.